Add files via upload
Commit based on issue

https://github.com/openssl/openssl/issues/17959

The example bits that worked were added. If I understand correctly it is not possible to do it "out of box" as x509_st is not clearly visible in /usr/include directory. Also, some useful function are static within the v3_purp.c file. The commit is mostly as a good discussion start point on the purpose modification issue.